requirements.txt
# ROHU Contable — Production Dependencies
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
Flask-Limiter==3.12
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
marshmallow==3.23.2
gunicorn==23.0.0
python-dotenv==1.0.1
argon2-cffi==23.1.0
pytesseract==0.3.13
Pillow==11.1.0
